AliView: DNA sequences viewer

This function uses AliView (Larsson, 2014) to visualize DNA sequences. The software must be installed on the computer.

aliview(
  x,
  aliview_exec = getOption("bioseq.aliview.exec", default = "aliview")
)

Arguments

x

a DNA, RNA or AA vector. Alternatively a DNAbin or AAbin object.

aliview_exec

a character string giving the path of the program.

Details

By default, the function assumes that the executable is installed in a directory located on the PATH. Alternatively the user can provide an absolute path to the executable (i.e. the location where the software was installed/uncompressed). This information can be stored in the global options settings using options(bioseq.aliview.exec = "my_path_to_aliview").

References

Larsson, A. (2014). AliView: a fast and lightweight alignment viewer and editor for large data sets. Bioinformatics 30(22): 3276-3278.

See also