Genetic code tables

List of all genetic code tables available in bioseq. The number in bold can be used to select a table in appropriate functions.

Available genetic codes

1. Standard

2. Vertebrate Mitochondrial

3. Yeast Mitochondrial

4. Mold Mitochondrial; Protozoan Mitochondrial; Coelenterate Mitochondrial; Mycoplasma; Spiroplasma

5. Invertebrate Mitochondrial

6. Ciliate Nuclear; Dasycladacean Nuclear; Hexamita Nuclear

9. Echinoderm Mitochondrial; Flatworm Mitochondrial

10. Euplotid Nuclear

11. Bacterial, Archaeal and Plant Plastid

12. Alternative Yeast Nuclear

13. Ascidian Mitochondrial

14. Alternative Flatworm Mitochondrial

15. Blepharisma Macronuclear

16. Chlorophycean Mitochondrial

21. Trematode Mitochondrial

22. Scenedesmus obliquus Mitochondrial

23. Thraustochytrium Mitochondrial

24. Pterobranchia Mitochondrial

25. Candidate Division SR1 and Gracilibacteria

26. Pachysolen tannophilus Nuclear

27. Karyorelict Nuclear

28. Condylostoma Nuclear

29. Mesodinium Nuclear

30. Peritrich Nuclear

31. Blastocrithidia Nuclear

32. Balanophoraceae Plastid

33. Cephalodiscidae Mitochondrial

References

Andrzej (Anjay) Elzanowski and Jim Ostell at National Center for Biotechnology Information (NCBI), Bethesda, Maryland, U.S.A. https://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=tgencodes