Convert DNAbin/AAbin to tibble

These methods convert sequences from ape formats DNAbin and AAbin to tibbles.

as_tibble.DNAbin(x, label = "label", sequence = "sequence", ...)

as_tibble.AAbin(x, label = "label", sequence = "sequence", ...)

Arguments

x
a DNAbin or AAbin object.

label
Name of the column that stores the sequence labels in the returned tibble.

sequence
Name of the column that stores the sequences in the returned tibble.

...
Not used.

Value

A tibble with two columns (if name is not NULL, the default) or one column (otherwise).

See also

Examples

require(ape)

#> Le chargement a nécessité le package : ape

require(tibble)

#> Le chargement a nécessité le package : tibble

x <- rDNAbin(nrow = 10, ncol = 25)
as_tibble.DNAbin(x)

#> # A tibble: 10 x 2
#>    label  sequence                 
#>    <chr>  <DNA>                    
#>  1 Ind_1  AGCCTAATGTGCGGAAACCATTCAC
#>  2 Ind_2  CACCATATTTGTTTAGTTGGTTATT
#>  3 Ind_3  TAGTGTCTGACAAACGCCTGCGCCT
#>  4 Ind_4  CACTAGGCATTACCACACGGTTCCT
#>  5 Ind_5  ACAGGCAGCAGACAGTTGAATCACC
#>  6 Ind_6  GTACCATCGCTAACCCATGAAAACT
#>  7 Ind_7  CCTAGGAGGTGGTATTCGACATCGT
#>  8 Ind_8  CAAGTGTCCGGACGCACAGGCACTT
#>  9 Ind_9  CGGCACTGAAAACATGATCACGCTT
#> 10 Ind_10 TGGCGGCCGGCATAGTGTTGAAACG